requirements.txt
flask
openai
line-bot-sdk
python-dotenv
gunicorn
psycopg2-binary
requests
transitions
graphviz
beautifulsoup4